pyproject.toml
[project]
name = "hackathon-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
